+++ server/controller/StrandController.js
@@ -1,4 +1,6 @@
 const Strand = require("../model/strands");
+const fs = require("fs");
+const path = require("path");
 
 class StrandController {
   /* CRUD ----------------------------------------------------------- */
@@ -6,6 +8,18 @@ class StrandController {
   async create(req, res) {
     const { strandTypeID, name, description } = req.body;
     const imagePath = req.file.path;
+
+    // Check if 'name' is missing
+    if (!name) {
+      throw new Error("Name field should be fill up!");
+    }
+
+    // Check if 'name' already exists in the database
+    const existingStrand = await Strand.findOne({ name });
+
+    if (existingStrand) {
+      throw new Error("This strand has already existed!");
+    }
 
     // INIT
     const newStrand = new Strand({
@@ -20,6 +34,18 @@ class StrandController {
 
     // RESPONSE
     res.json({ strandType: strandTypeID, name, description, imagePath });
+  }
+
+  // AUTH
+  async auth(req, res) {
+    res.json({
+      user: req.user,
+      selectedStrand: req.selectedStrand,
+      preferredStrand: req.preferredStrand,
+      personalEngagements: req.pes,
+      subjects: req.subjects,
+      pendingSubjects: req.pendingSubjects,
+    });
   }
 
   // READ ALL
@@ -67,11 +93,24 @@ class StrandController {
   async delete(req, res) {
     const { strandID } = req.params;
 
+    // Find the Strand data first to get the imagePath
+    const strand = await Strand.findById(strandID);
+
+    if (!strand) {
+      throw new Error("Strand not found.");
+    }
+
+    // DELETE IMAGE
+    if (strand.imagePath) {
+      const imagePath = path.join(__dirname, "../", strand.imagePath);
+      fs.unlinkSync(imagePath);
+    }
+
     // DELETE SINGLE DATA
-    const strand = await Strand.deleteOne({ _id: strandID });
+    const deletedStrand = await Strand.deleteOne({ _id: strandID });
 
     // RESPONSE
-    res.json(strand);
+    res.json(deletedStrand);
   }
 
   // DELETE ALL
